Getting data, testing, scoring
trainTestSet: attaches header information in numbers for use in sklearn. Included are the seconds since 5 (either am or pm), whether it's evening and the day of the week (Monday=0...Friday=4). The train is assigned a color number based on a subjective interpretation of how 'far' train lines are from each other, e.g., Silver is close to Orange but far from Red. It is also notes whether or not the train is headed toward the north/east or south/west. Last, it adds the information for the previous train, as that seems to have predictive value.
trainDataCSV: load the data from a CSV file and make it consistent with the other data
trainData: load train data from the SQL server. For instance,
BYtrainsAugNE=trainData(BYLine,['G05','E01','E06','E10'],8,[1,2,3,4,5,8,9,10,11,12])
loads all the data for all the Blue and Yellow trains from King St to Pentagon toward Largo (G05), Mt. Vernon (E01), etc.
scoreModel: a quick way to score the models

For example,
GRsurge2trainsS=trainDat